requirements.txt
amocrm-api==0.8.3
certifi==2017.11.5
chardet==3.0.4
cookies==2.2.1
Django==2.0
djangorestframework==3.7.3
gunicorn==19.7.1
idna==2.6
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
responses==0.8.1
six==1.11.0
urllib3==1.22
